Add initial AddEditBookActivity abstract class
